do not load BSP tree into RAM
These changes, along with changes in the walls3d program, free RAM and allow for larger maps by leaving the BSP tree in flash and not loading it as a whole into RAM. This involves a change to the serialized tree format in which child node indices are added.

In-flash tree traversal does come at the cost of a slightly lower frame rate (e.g. 126 ms/frame vs. 109 ms/frame for initial screen in the "smiley face" map). This could potentially be ameliorated by having a caching system by which some nodes are kept in RAM.